Basic part finished

+++ PAS-shopApp/src/main/java/com/pas/manager/OrderManager.java
@@ -8,6 +8,7 @@ import com.pas.repository.OrderRepository;
 import com.pas.repository.UserRepository;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.persistence.EntityNotFoundException;
 
 import java.util.*;
 import java.util.concurrent.atomic.AtomicReference;
@@ -25,57 +26,67 @@ public class OrderManager {
         return orderRepository.findById(id);
     }
 
-    public List<Order> findAllOrders(){
+    public List<Order> findAllOrders() {
         return orderRepository.findAll();
     }
 
     public Order createOrder(UUID userId, Address shippingAddress) {
         Optional<User> customer = userRepository.findById(userId);
-        Map<Product,Long> orderItems = customer.get().getCart().getItems();
+        Map<Product, Long> orderItems = customer.get().getCart().getItems();
         Double orderValue = calculateOrderValue(orderItems);
 
-        if( customer.isPresent() && shouldCreateOrder(userId, orderItems, shippingAddress, orderValue)){
+        if (customer.isPresent() && shouldCreateOrder(userId, orderItems, shippingAddress, orderValue)) {
             process(customer.get(), orderItems, orderValue);
-           return orderRepository.add(new Order(customer.get(), shippingAddress, orderItems, new Date(), true, 0, false));
+            Order order = orderRepository.add(new Order(customer.get(), shippingAddress, orderItems, new Date(), true, 0, false));
+            clearUserCart(customer.get());
+            return order;
         } else {
             throw new IllegalStateException("violated business logic");
         }
     }
 
-    public void cancelOrder(UUID id) { orderRepository.delete(id);}
+    public void cancelOrder(UUID id) {
+        orderRepository.delete(id);
+    }
 
-    public void deliverOrder(UUID orderId){
+    public void deliverOrder(UUID orderId) {
         Optional<Order> found = orderRepository.findById(orderId);
-        if(found.isPresent()){
+        if (found.isPresent()) {
             found.get().setDelivered(true);
+            found.get().setDeliveryDate(new Date());
             orderRepository.update(found.get().getId(), found.get());
         }
     }
 
-    private void process(User user, Map<Product, Long> products, Double orderValue){
+    private void process(User user, Map<Product, Long> products, Double orderValue) {
         user.setAccountBalance(user.getAccountBalance() - orderValue);
-        products.entrySet().forEach(key ->{
+        products.entrySet().forEach(key -> {
             key.getKey().setAvailableAmount(key.getKey().getAvailableAmount() - key.getValue().intValue());
         });
     }
 
-    private boolean shouldCreateOrder(UUID userId, Map<Product,Long> orderItems, Address shippingAddress, Double orderValue){
-        if( isEnoughItems(orderItems) && !isUserSuspended(userId) && checkIfEnoughMoney(userId, orderValue)){
+    private void clearUserCart(User user) {
+        user.getCart().setItems(new HashMap<>());
+        userRepository.update(user.getId(), user);
+    }
+
+    private boolean shouldCreateOrder(UUID userId, Map<Product, Long> orderItems, Address shippingAddress, Double orderValue) {
+        if (isEnoughItems(orderItems) && !isUserSuspended(userId) && checkIfEnoughMoney(userId, orderValue)) {
             return true;
         } else {
-             return false;
+            return false;
         }
     }
 
-    private boolean isEnoughItems(Map<Product,Long> orderItems){
+    private boolean isEnoughItems(Map<Product, Long> orderItems) {
         return orderItems.entrySet().stream()
-                .map(( entry -> isEnoughItems(entry.getKey(), entry.getValue())))
-                .filter( result -> result.equals(true))
+                .map((entry -> isEnoughItems(entry.getKey(), entry.getValue())))
+                .filter(result -> result.equals(true))
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("Cant create order, items out of stock"));
     }
 
-    private boolean isUserSuspended(UUID userId){
+    private boolean isUserSuspended(UUID userId) {
         User found = userRepository.findById(userId).orElse(null);
         return Optional.ofNullable(found)
                 .map(User::isSuspended)
@@ -83,22 +94,23 @@ public class OrderManager {
                 .orElseThrow(() -> new IllegalStateException("Cant create order, customer suspended"));
     }
 
-    private boolean checkIfEnoughMoney(UUID userId, Double orderValue){
+    private boolean checkIfEnoughMoney(UUID userId, Double orderValue) {
         User found = userRepository.findById(userId).orElse(null);
         return Optional.ofNullable(found)
                 .map(user -> isEnoughMoney(user.getAccountBalance(), orderValue))
                 .filter(result -> result.equals(true))
                 .orElseThrow(() -> new IllegalStateException("Cant create order, user don't have enough money"));
     }
-    private boolean isEnoughItems(Product product, Long availableAmount){
+
+    private boolean isEnoughItems(Product product, Long availableAmount) {
         return product.getAvailableAmount() - availableAmount >= 0;
     }
 
-    private boolean isEnoughMoney(Double userMoney, Double orderValue){
-        return userMoney -  orderValue >= 0;
+    private boolean isEnoughMoney(Double userMoney, Double orderValue) {
+        return userMoney - orderValue >= 0;
     }
 
-    private Double calculateOrderValue(Map<Product,Long> orderItems){
+    private Double calculateOrderValue(Map<Product, Long> orderItems) {
         AtomicReference<Double> orderValue = new AtomicReference<>((double) 0);
         orderItems.entrySet().forEach(key -> {
             orderValue.updateAndGet(v -> v + key.getKey().getPrice() * key.getValue());
@@ -107,13 +119,23 @@ public class OrderManager {
     }
 
     public void deleteOrder(UUID orderId) {
-        orderRepository.delete(orderId);
+        User user = findUserInOrder(orderId);
+        Order order = orderRepository.findById(orderId).orElseThrow(() -> new EntityNotFoundException("entity dont exist"));
+        if(order.isDelivered()) {
+            user.setAccountBalance(user.getAccountBalance() + calculateOrderValue(order.getItems()));
+            orderRepository.delete(orderId);
+        }
     }
-    public List<Order> findOngoingOrders(){
+
+    public List<Order> findOngoingOrders() {
         return orderRepository.findOngoingOrders();
     }
 
-    public List<Order> findFinishedOrders(){
+    public List<Order> findFinishedOrders() {
         return orderRepository.findFinishedOrders();
+    }
+
+    public User findUserInOrder(UUID orderId){
+        return orderRepository.findById(orderId).orElseThrow(() -> new EntityNotFoundException("entity dont exist")).getCustomer();
     }
 }
